+++ holoso/_frontend/__init__.py
@@ -1,4 +1,3 @@
-"""Front end: lower a Python function object into HIR, plus the shared output-port naming convention."""
+"""Front end: lower a Python function object into HIR."""
 
 from ._lower import lower as lower
-from ._shape import flatten_value as flatten_value, port_name as port_name
