--- tests/test_shape.py
@@ -1,58 +1,0 @@
-"""Unit tests for the frontend output-flattening / port-naming convention."""
-
-import dataclasses
-
-from holoso._frontend import flatten_value, port_name
-
-from ._modelref import output_names
-
-
-def test_scalar_is_output_zero() -> None:
-    assert output_names(3.14) == ["out_0"]
-
-
-def test_flat_sequence_is_positional() -> None:
-    assert output_names((1.0, 2.0, 3.0)) == ["out_0", "out_1", "out_2"]
-
-
-def test_nested_list_row_major_like_ekf1() -> None:
-    # ekf1's update_x_P returns a 9x1 nested list -> out_0_0 .. out_8_0
-    matrix = [[float(i)] for i in range(9)]
-    assert output_names(matrix) == [f"out_{i}_0" for i in range(9)]
-
-
-def test_matrix_n_by_m() -> None:
-    matrix = [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]]
-    assert output_names(matrix) == ["out_0_0", "out_0_1", "out_0_2", "out_1_0", "out_1_1", "out_1_2"]
-
-
-def test_dataclass_fields_and_nesting() -> None:
-    @dataclasses.dataclass
-    class Foo:
-        bar: float
-
-    @dataclasses.dataclass
-    class Baz:
-        foo: Foo
-
-    assert output_names((Baz(Foo(1.0)), 2.0)) == ["out_0_foo_bar", "out_1"]
-
-
-def test_bare_dataclass_uses_field_names() -> None:
-    @dataclasses.dataclass
-    class Out:
-        x: float
-        y: float
-
-    assert output_names(Out(1.0, 2.0)) == ["out_x", "out_y"]
-
-
-def test_port_name_paths() -> None:
-    assert port_name((0,)) == "out_0"
-    assert port_name((0, "foo", "bar")) == "out_0_foo_bar"
-    assert port_name((3, 1)) == "out_3_1"
-
-
-def test_flatten_value_returns_leaves() -> None:
-    leaves = flatten_value([[1.5], [2.5]])
-    assert [value for _, value in leaves] == [1.5, 2.5]
